Favorite Recipes Display
   - Display the favorite recipes saved in the DB using the FavoritesRecipesFragment.kt
   - Refactor the FavoritesRecipesFragment.kt class name
