+++ app/src/main/res/layout/fragment_favorites_receipes.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/lightGray"
-    tools:context=".ui.fragments.FavoritesReceipesFragment">
+    tools:context=".ui.fragments.FavoritesRecipesFragment">
 
 
     <androidx.recyclerview.widget.RecyclerView
